+++ packages/native/src/components/Text/Text.tsx
@@ -1,10 +1,5 @@
 import React from 'react';
-import { StyleProp, TextStyle as RNTextStyle } from 'react-native';
 
-import {
-  TextBaseProps,
-  TextTypes,
-} from '@a-little-world/little-world-design-system-core';
 import BaseText, { BaseTextProps } from './BaseText';
 import textParser from '../../utils/parser';